+++ src/voip/msiframerequestslimiter.c
@@ -30,7 +30,7 @@ void ms_iframe_requests_limiter_request_iframe(MSIFrameRequestsLimiterCtx *obj) 
 }
 
 bool_t ms_iframe_requests_limiter_iframe_requested(const MSIFrameRequestsLimiterCtx *obj, uint64_t curtime) {
-	return obj->iframe_required && ( obj->last_sent_iframe_time == (uint64_t)-1 || (curtime - obj->last_sent_iframe_time) > obj->min_iframe_interval);
+	return obj->iframe_required && ( obj->last_sent_iframe_time == (uint64_t)-1 || (int)(curtime - obj->last_sent_iframe_time) > obj->min_iframe_interval);
 }
 
 void ms_iframe_requests_limiter_notify_iframe_sent(MSIFrameRequestsLimiterCtx *obj, uint64_t curtime) {
